conversation-manager/requirements.txt
gunicorn
flask
Flask-SQLAlchemy
Flask-API
PyMySQL
Werkzeug==2.3.0
minio
requests
pyyaml